+++ node_interop/lib/os.dart
@@ -12,7 +12,7 @@ import 'package:js/js.dart';
 import 'node.dart';
 
 OS get os => _os ??= require('os');
-OS _os;
+OS? _os;
 
 @JS()
 @anonymous