+++ node_interop/lib/path.dart
@@ -12,7 +12,7 @@ import 'package:js/js.dart';
 import 'node.dart';
 
 Path get path => _path ??= require('path');
-Path _path;
+Path? _path;
 
 @JS()
 @anonymous