+++ node_io/test/internet_address_test.dart
@@ -56,9 +56,10 @@ void main() {
     });
 
     test('lookup successful', () async {
-      var list = await InternetAddress.lookup('google.com');
-      expect(list, isNotEmpty);
-      expect(list.first.type, equals(InternetAddressType.IPv4));
+      var list = (await InternetAddress.lookup('google.com'))
+          .map((address) => address.type);
+      expect(list.contains(InternetAddressType.IPv4), equals(true));
+      expect(list.contains(InternetAddressType.IPv6), equals(true));
     });
 
     test('reverse successful', () async {