+++ build_node_compilers/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.3.2
+
+- Allowed Dart2JS to run in unsound null-safety mode.
+- Removed dependency on deprecated analyzer packages.
+
 # 0.3.1
 
 - Upgraded dependency on build_modules to `2.11.3` to fix issues with Dart SDK constraint.